Implement new command-line options and which option takes precedent
When giving command-line options as well as the config option the
respective values in the config file were supposed to be ignored.
There was no option to speciffically say unordered/sets, it was just
assumed if no option was given for ordered/bags then it must be
unordered/sets. This meant that the config file values would
always be ignored.

* Add options "unord" and "set"
* Implement importance for each option - config option least
  "important", ord option more "important" than unord etc
* Add a "\settings" command in the interpreter that allows the
  user to view the configuration - ordered/unordered and bags/sets